+++ models/all/src/index.ts
@@ -66,7 +66,6 @@ const builders: [(b: Builder) => void, string][] = [
   [activityModel, 'activity'],
   [attachmentModel, 'attachment'],
   [tagsModel, 'tags'],
-  [calendarModel, 'calendar'],
   [viewModel, 'view'],
   [workbenchModel, 'workbench'],
   [contactModel, 'contact'],
@@ -101,7 +100,8 @@ const builders: [(b: Builder) => void, string][] = [
   [serverGmailModel, 'server-gmail'],
   [serverTelegramModel, 'server-telegram'],
   [trackerModel, 'tracker'],
-  [boardModel, 'board']
+  [boardModel, 'board'],
+  [calendarModel, 'calendar']
 ]
 
 for (const [b, id] of builders) {