+++ changelog.md
@@ -13,6 +13,7 @@ Leads:
 Tracker:
 
 - Attachments support
+- Board view
 
 ## 0.6.26